+++ Src/SOS.Observations.Api/Controllers/AreasController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using SOS.Lib.Models.Search;
+using SOS.Lib.Models.Shared;
 using SOS.Observations.Api.Controllers.Interfaces;
 using SOS.Observations.Api.Dtos;
 using SOS.Observations.Api.Dtos.Enum;
@@ -59,36 +60,11 @@ namespace SOS.Observations.Api.Controllers
         [ProducesResponseType(typeof(byte[]), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        public async Task<IActionResult> GetExport([FromRoute] AreaTypeDto areaType, [FromRoute] string featureId)
+        public async Task<IActionResult> GetExport([FromRoute] AreaTypeDto areaType, [FromRoute] string featureId, [FromQuery] AreaExportFormatDto format = AreaExportFormatDto.Json)
         {
             try
             {
-                var zipBytes = await _areaManager.GetZippedAreaAsync(areaType, featureId);
-
-                if (zipBytes == null)
-                {
-                    return new StatusCodeResult((int)HttpStatusCode.NoContent);
-                }
-
-                return File(zipBytes, "application/zip", $"Area{areaType:G}:{featureId}.zip");
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "Error getting areas");
-                return new StatusCodeResult((int)HttpStatusCode.InternalServerError);
-            }
-        }
-
-        /// <inheritdoc />
-        [HttpGet("{areaType}/{featureId}/ExportGeoJson")]
-        [ProducesResponseType(typeof(byte[]), (int)HttpStatusCode.OK)]
-        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        public async Task<IActionResult> GetExportGeoJson([FromRoute] AreaTypeDto areaType, [FromRoute] string featureId)
-        {
-            try
-            {
-                var zipBytes = await _areaManager.GetZippedAreaGeoJsonAsync(areaType, featureId);
+                var zipBytes = await _areaManager.GetZippedAreaAsync(areaType, featureId,(AreaExportFormat) format);
 
                 if (zipBytes == null)
                 {