+++ Src/SOS.Observations.Api/Dtos/AreaExportFormatDto.cs
@@ -1,0 +1,23 @@
+﻿namespace SOS.Observations.Api.Dtos
+{
+    /// <summary>
+    /// Export format for area.
+    /// </summary>
+    public enum AreaExportFormatDto
+    {
+        /// <summary>
+        /// JSON.
+        /// </summary>
+        Json,
+
+        /// <summary>
+        /// GeoJSON.
+        /// </summary>
+        GeoJson,
+
+        /// <summary>
+        /// WKT.
+        /// </summary>
+        Wkt
+    }
+}