+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - New search filter: `BirdNestActivityLimit`
 - New vocabulary: `BirdNestActivity`
 - Possibility to get result as GeoJSON in the following endpoint: `/Observations/Internal/Search`
+- Possibility to export areas as JSON, GeoJSON or WKT in the `/Areas/{areaType}/{fetureId}` endpoint.
 
 ### `Changed`
 - Geting a single observation by occurrence id (both external and internal) return one observation and not an array with one observation